FileIOchooser split to LoadChooser and SaveChooser
