+++ src/org/aavso/tools/vstar/ui/mediator/Mediator.java
@@ -66,6 +66,7 @@ import org.aavso.tools.vstar.ui.dialog.DelimitedFieldFileSaveChooser;
 import org.aavso.tools.vstar.ui.dialog.DiscrepantReportDialog;
 import org.aavso.tools.vstar.ui.dialog.DoubleField;
 import org.aavso.tools.vstar.ui.dialog.IntegerField;
+import org.aavso.tools.vstar.ui.dialog.LoadChooser;
 import org.aavso.tools.vstar.ui.dialog.MessageBox;
 import org.aavso.tools.vstar.ui.dialog.MultiEntryComponentDialog;
 import org.aavso.tools.vstar.ui.dialog.ObservationDetailsDialog;
@@ -73,8 +74,7 @@ import org.aavso.tools.vstar.ui.dialog.PNGImageFileSaveChooser;
 import org.aavso.tools.vstar.ui.dialog.PhaseDialog;
 import org.aavso.tools.vstar.ui.dialog.PhaseParameterDialog;
 import org.aavso.tools.vstar.ui.dialog.PlotControlDialog;
-import org.aavso.tools.vstar.ui.dialog.FileIOchoosers;
-import org.aavso.tools.vstar.ui.dialog.FileIOchoosers.fileIOchooserMode;
+import org.aavso.tools.vstar.ui.dialog.SaveChooser;
 import org.aavso.tools.vstar.ui.dialog.filter.ObservationFilterDialog;
 import org.aavso.tools.vstar.ui.dialog.filter.ObservationFiltersDialog;
 import org.aavso.tools.vstar.ui.dialog.model.ModelDialog;
@@ -224,9 +224,12 @@ public class Mediator {
 	// A file dialog for saving a VeLa code file.
 	private VeLaFileSaveChooser velaFileSaveDialog;
 	
-	// Methods to load/save an XML as String from/to VeLa Model XML files with respective choosers.  
-	private FileIOchoosers velaXMLchoosers;
+	// A helper object to load an XML from VeLa Model XML file using respective chooser.  
+	private LoadChooser velaXMLloadDialog;
 
+	// An helper object to save a VeLa model to XML file using respective chooser.  
+	private SaveChooser velaXMLsaveDialog;
+	
 	// Persistent phase parameter dialog.
 	private PhaseParameterDialog phaseParameterDialog;
 
@@ -321,9 +324,8 @@ public class Mediator {
 			extensionFilterOpen[0] = new FileNameExtensionFilter("VeLa XML files (*.vlx)", "vlx");
 			extensionFilterOpen[1] = new FileNameExtensionFilter("VeLa files (*.txt, *.vl, *.vela)", "txt", "vl", "vela");
 			extensionFilterSave[0] = extensionFilterOpen[0];
-			this.velaXMLchoosers = new FileIOchoosers(extensionFilterOpen, extensionFilterSave, 
-					"vlx", "Open VeLa XML File", "Save VeLa XML File As",
-					fileIOchooserMode.OPEN_SAVE);
+			this.velaXMLloadDialog = new LoadChooser(extensionFilterOpen, "Open VeLa XML File");
+			this.velaXMLsaveDialog = new SaveChooser(extensionFilterSave, "vlx", "Save VeLa XML File");
 		}
 
 		// These (among other things) are created for each new star.
@@ -612,12 +614,19 @@ public class Mediator {
 	}
 	
 	/**
-	 * @return the velaXMLchoosers
+	 * @return the velaXMLloadDialog
 	 */
-	public FileIOchoosers getVelaXMLchoosers() {
-		return velaXMLchoosers;
+	public LoadChooser getVelaXMLloadDialog() {
+		return velaXMLloadDialog;
 	}
 
+	/**
+	 * @return the velaXMLsaveDialog
+	 */
+	public SaveChooser getVelaXMLsaveDialog() {
+		return velaXMLsaveDialog;
+	}
+	
 	/**
 	 * Create a mean observation change listener and return it. Whenever the mean
 	 * series source changes, listeners may want to perform a new period analysis or
